refactor(ui): split TypingTestView into stats and controls rows
